+++ src/constants/networks.ts
@@ -1,5 +1,5 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { ApolloClient, InMemoryCache, NormalizedCacheObject } from '@apollo/client'
+import { NetworkInfo } from './type'
 
 import ethereumInfo from './networks/ethereum'
 import ropstenInfo from './networks/ropsten'
@@ -22,12 +22,6 @@ import velasInfo from './networks/velas'
 import auroraInfo from './networks/aurora'
 import oasisInfo from './networks/oasis'
 import optimismInfo from './networks/optimism'
-
-export const createClient = (url: string): ApolloClient<NormalizedCacheObject> =>
-  new ApolloClient({
-    uri: url,
-    cache: new InMemoryCache(),
-  })
 
 export const SUPPORTED_NETWORKS: ChainId[] = [
   ChainId.MAINNET,
@@ -65,64 +59,6 @@ export const TRUESIGHT_NETWORK_TO_CHAINID: NetToChain = {
   polygon: ChainId.MATIC,
   fantom: ChainId.FANTOM,
   cronos: ChainId.CRONOS,
-}
-
-export type NetworkInfo = {
-  readonly chainId: ChainId
-  readonly route: string
-  readonly name: string
-  readonly icon: string
-  readonly classicClient: ApolloClient<NormalizedCacheObject>
-  readonly elasticClient: ApolloClient<NormalizedCacheObject>
-  readonly blockClient: ApolloClient<NormalizedCacheObject>
-  readonly etherscanUrl: string
-  readonly etherscanName: string
-  readonly tokenListUrl: string
-  readonly bridgeURL: string
-  readonly nativeToken: {
-    readonly symbol: string
-    readonly name: string
-    readonly address: string
-    readonly logo: string
-    readonly decimal: number
-  }
-  readonly rpcUrl: string
-  readonly routerUri: string
-  readonly classic: {
-    readonly static: {
-      readonly zap: string
-      readonly router: string
-      readonly factory: string
-    }
-    readonly dynamic: {
-      readonly zap: string
-      readonly router: string
-      readonly factory: string
-    } | null
-    readonly routerV2: string //todo: remove in future
-    readonly aggregationExecutor: string //todo: remove in future
-    readonly factory: string
-    readonly migrate: string
-    readonly claimReward: string
-    readonly fairlaunch: string[]
-    readonly fairlaunchV2: string[]
-  }
-  readonly elastic: {
-    readonly coreFactory: string
-    readonly nonfungiblePositionManager: string
-    readonly tickReader: string
-    readonly initCodeHash: string
-    readonly quoter: string
-    readonly routers: string
-  }
-  // token: {
-  //   DAI: Token
-  //   USDC: Token
-  //   USDT: Token
-  // }
-  readonly avgrageBlockTimeInSeconds: number
-  readonly coingeckoNetworkId: string //https://api.coingecko.com/api/v3/asset_platforms
-  readonly coingeckoNativeTokenId: string //https://api.coingecko.com/api/v3/coins/list
 }
 
 const NETWORKS_INFO_CONFIG: { [chain in ChainId]: NetworkInfo } = {
