+++ src/constants/networks/avax-testnet.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import AVAX from '../assets/networks/avax-network.png'
+import AVAX from 'assets/networks/avax-network.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -40,7 +41,6 @@ const avaxTestnetInfo: NetworkInfo = {
     },
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0x7900309d0b1c8D3d665Ae40e712E8ba4FC4F5453',
     migrate: EMPTY,
     claimReward: EMPTY,
     fairlaunch: ['0xC3E2aED41ECdFB1ad41ED20D45377Da98D5489dD'],
