+++ src/constants/networks/bnb.ts
@@ -1,8 +1,9 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import BSC from '../assets/networks/bsc-network.png'
-import BnbLogo from '../assets/images/bnb-logo.png'
+import BSC from 'assets/networks/bsc-network.png'
+import BnbLogo from 'assets/images/bnb-logo.png'
 
 const EMPTY = ''
 
@@ -40,7 +41,6 @@ const bnbInfo: NetworkInfo = {
     },
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0x878dFE971d44e9122048308301F540910Bbd934c',
     migrate: EMPTY,
     claimReward: EMPTY,
     fairlaunch: [
