+++ src/constants/networks/bttc.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import BTT from '../assets/networks/bttc.png'
+import BTT from 'assets/networks/bttc.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -44,7 +45,6 @@ const bttcInfo: NetworkInfo = {
     },
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0xD9bfE9979e9CA4b2fe84bA5d4Cf963bBcB376974',
     migrate: EMPTY,
     claimReward: '0x1a91f5ADc7cB5763d35A26e98A18520CB9b67e70',
     fairlaunch: EMPTY_ARRAY,
