+++ src/constants/networks/fantom.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import FTM from '../assets/networks/fantom-network.png'
+import FTM from 'assets/networks/fantom-network.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -41,7 +42,6 @@ const fantomInfo: NetworkInfo = {
     },
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0x78df70615ffc8066cc0887917f2Cd72092C86409',
     migrate: EMPTY,
     claimReward: EMPTY,
     fairlaunch: EMPTY_ARRAY,
