+++ src/constants/networks/matic.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import Polygon from '../assets/networks/polygon-network.png'
+import Polygon from 'assets/networks/polygon-network.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -40,7 +41,6 @@ const maticInfo: NetworkInfo = {
     },
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0x5F1fe642060B5B9658C15721Ea22E982643c095c',
     migrate: EMPTY,
     claimReward: '0x89929Bc485cE72D2Af7b7283B40b921e9F4f80b3',
     fairlaunch: [
