+++ src/constants/networks/optimism.ts
@@ -1,8 +1,9 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import OPTIMISM from '../assets/networks/optimism-network.svg'
-import EthereumLogo from '../assets/images/ethereum-logo.png'
+import OPTIMISM from 'assets/networks/optimism-network.svg'
+import EthereumLogo from 'assets/images/ethereum-logo.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -40,7 +41,6 @@ const optimismInfo: NetworkInfo = {
     dynamic: null && 'not support',
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0x1c758aF0688502e49140230F6b0EBd376d429be5',
     migrate: EMPTY,
     claimReward: EMPTY,
     fairlaunch: EMPTY_ARRAY,
