+++ src/constants/networks/velas.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { createClient, NetworkInfo } from 'constants/networks'
+import { NetworkInfo } from '../type'
+import { createClient } from 'utils/client'
 
-import VELAS from '../assets/networks/velas-network.png'
+import VELAS from 'assets/networks/velas-network.png'
 
 const EMPTY = ''
 const EMPTY_ARRAY: any[] = []
@@ -40,7 +41,6 @@ const velasInfo: NetworkInfo = {
     dynamic: null && 'not support',
     routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
     aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
-    factory: '0xD9bfE9979e9CA4b2fe84bA5d4Cf963bBcB376974',
     migrate: 'https://bridge.velaspad.io',
     claimReward: EMPTY,
     fairlaunch: EMPTY_ARRAY,
