+++ src/utils/client.ts
@@ -1,0 +1,7 @@
+import { ApolloClient, InMemoryCache, NormalizedCacheObject } from '@apollo/client'
+
+export const createClient = (url: string): ApolloClient<NormalizedCacheObject> =>
+  new ApolloClient({
+    uri: url,
+    cache: new InMemoryCache(),
+  })
